Add PPC440 in Virtex-5 FXT to CPU table
Hi Wolfgang,

attach find a patch that adds the PPC440 in Virtex-5 FXT to the list of
known CPUs in the kernel. This patch is important as without it the
Linux kernel does not boot when the data cache is enabled due to
incorrect CPU features.

Please apply to the GIT tree.

Thanks,
- Peter