[XILINX][LLTEMAC] Fixed to work on ML405 board.
This required *not* defining the special PHY programming code,
which apparently doesn't work, and adding the xenv.h header
file, which was never copied.  This header file contains some
macros for flushing the cache buffers which are used by the
DMA code.  Without these defines, the code defaults to a noop,
and the DMA issues an error status, which appears to be
unrecoverable.

Also added a couple of minor cleanups and some better
debugging messages.