Xilinx Microblaze: Fix cache routines to use cpuinfo.
Moving away from xparameters defines for cache sizes and line sizes.
This introduces some minimal additional overhead to query the cpuinfo
struct.  However, compared to the amount of code already there, this
is minimal.  Removed the LINE_SIZE defines from cache.h, since they
are no longer needed.

Note that for cacheless systems, there is some additional runtime
overhead that was not there before, since all the cache management
code was #ifdef'd away.  I believe that this should not cause a
huge burden for cacheless systems.  If this is a concern, then I
suggest adding a config option which would eliminate all the cache
management code.  This option would default to including the code,
but could be activated as an optimization.