Xilinx Microblaze: cache.c no longer depends on xparameters.
The last thing remaining that depends on xparameters was a check for
whether cache invalidations have to be aligned.  Based on the previous
code, this is required for FSL memory interfaces, but not for PLB
memory interfaces.  Since PLB memory interfaces are no longer used in
current microblazes, we'll just assume that this alignment is always
required.