Xilinx: hwicap: refactoring and support for fifo-based icap.
This includes code for new fifo-based xps_hwicap which has a
significantly different interface from the old opb_hwicap.  This
includes a fairly significant refactoring of the old icap code,
intended to share as much of the 'uninteresting' code between the two
drivers as possible.  The driver for the fifo_icap only supports
polled mode, in order to make this driver simpler, and since the
interrupt processing overhead is likely to slow down the throughput of
the driver until the core is updated to be a master on the bus.

One of the key refactorings includes dealing with the fact that the
ICAP has a slightly different set of registers in each architecture.
Previously this was handled as a config option.  This is not-so-nice
for ARCH=powerpc, since we'd like to have multiplatform support.  So,
at least for the of_driver advertisement, this is now selected based
on information from the device tree.  For platform devices which are
likely only used in ARCH=ppc, this still relies on a CONFIG option.

Signed-off-by: Stephen Neuendorffer <stephen.neuendorffer@xilinx.com>