Xilinx: FPU: Workaround for program exception with FPU after lazy context switch
When the FPU IP is connected to the APU controller of the Xilinx PPC440
and the FPU is enabled the processor can erroneously take a program
exception instead of an FPU unavailable exception when it hits the first
FPU instruction after a lazy context switch.

This patch works around the problem.