Xilinx: Boot: Fix bootstrap loader for 16550 UART
The virtex.c did not open the serial console before doing a printf
such that it would not work with the UART 16550. Console init does
not completely initialize the uart without opening it. It worked with
the UART lite because it requires no setup like the 16550.

The 1st version of this patch did not cover the case for when stdout
was not defined, so the 2nd version checks to make sure the open
function is defined.

Signed-off-by: John Linn <john.linn@xilinx.com>