Xilinx: Updates for future ML5 merge into main branch
An early merge revealed a couple problems when merging.
The first issue was a new #include of xparameters.h in
virtex.h which broke the console with the 405. The

The 2nd problem was in virtex_devices.c with the platform
data and dcr_host. This was changed to be set based on
the DCR xparameter so that it would work for 405 and 440.
The DCR config was also removed from the TEMAC Kconfig
as the DCR code can be built all the time.

Signed-off-by: John Linn <john.linn@xilinx.com>