Xilinx: Support for compressed device trees in BRAM
Added boot code which assumes that a compressed device tree blob
exists in BRAM.  The code decompresses it next to the kernel before
passing it to the kernel, enabling the device tree to be tightly bound
to a bitstream, rather than being bound to the kernel.  This makes it
easy to build a single kernel and run it on many different boards.

Signed-off-by: Stephen Neuendorffer <stephen.neuendorffer@xilinx.com>