[POWERPC] Xilinx: add compatibility for 'simple-bus'.
ePAPR drafts propose 'simple-bus' as a generic compatibility type for
busses which cannot be probed for devices.  In addition, the Xilinx
versions of these IPs seem to be proliferating.  Hence, in the future
let's prefer to use the standard names.  I've left the old names in
for short term backward compatibility for existing device trees.

Signed-off-by: Stephen Neuendorffer <stephen.neuendorffer@xilinx.com>