Xilinx: Don't use PPC_UDBG_16550 with Virtex platforms.
The Xilinx 16550 serial port IP cores require reg-shift and reg-offset
properties, which are not supported by the legacy_serial driver.  This
patch removes support for the legacy_serial driver on this platform,
in order to prevent it from mathing in the device tree.  This problem
was recently identified with device trees that contain 'simple-bus'
compatibility.

Signed-off-by: Stephen Neuendorffer <stephen.neuendorffer@xilinx.com>
Signed-off-by: John Linn <john.linn@xilinx.com>