Powerpc: LL TEMAC: Update for 2.6.30 crash with platform device
A change was made in 2.6.30, referred to below, which broke the
driver and there was no longer any need to support the normal
device driver registration model since we're using the device tree
(OF) model now. All the non-OF driver registration code was removed
to simplify the driver.

Since Grant's new flat driver is coming there's no need to invest
a lot in this driver.

commit: 57fee4a58fe802272742caae248872c392a60670.
platform: introduce module id table for platform devicesA

Signed-off-by: John Linn <john.linn@xilinx.com>