spi: Fix "not ready" timeout after write in at25.
Under certain circumstances msleep(1) within the loop, which waits
for the EEPROM to be finished, might take longer than the timeout.
On the next loop the status register might now return to be ready
and therefore the loop finishes. The following check now tests if
a timeout occurred and if so returns an error although the device
reported it was ready.

This fix replaces testing the occurrence of the timeout by testing
the "not ready" bit in the status register.

Signed-off-by: Sebastian Heutling <heutling@who-ing.de>
Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>