Xilinx: ARM: Creation of new Pele branch based on Dragonfire branch.
This commit is the 1st commit which is a basic kernel without any
drivers. It is intended to run on EP 4.1.2 with the ARM 1176. Since
there's no serial driver yet it won't run easily on the h/w.

It is a 2.6.32 kernel from kernel.org with the core kernel/BSP
changes from the Dragonfire branch applied to it.

This does not include all DF drivers yet on purpose with the goal
of making it easier to transition to a new kernel in the future.