Xilinx: ARM: xilinx_defconfig update for A9 hacks
This commit should be able to be backed out as it has 2 changes in it
that are a result of other commits. It has the 32v6K feature turned
off and the HZ slower. These two are both a result of the Kconfigs
being changed (arch/arm/Kconfig and arch/arm/mm/Kconfig) in previous
commits.

The 32v6K change was part of the LDREX/STREX problem and this was an
easy way to take out part of them (1st try).