Xilinx: ARM: Cleanup h/w addresses, added SCU mmu entry
A lot of address cleanup as it's messy with the old PSS GIC and TTC
being used til the new ones are tested. A new mmu entry was added
for the SCU so that it can be accessed.

DMA was removed from the makefile as it was causing the kernel to
crash with the new addresses. More investigation needed for DMA to
be put back in.