Xilinx: ARM: Adding smp files
Added SMP files so that infrastructure can be moving in that direction
as some of the SCU stuff, like timers, is needed prior to real SMP.

There's still a lot of work to be done as functions are stubbed out
or not changed at all. They were copied from the omap kernel.

This code won't be enabled and built into the kernel until SMP is
enabled in the kernel config under kernel features.