Xilinx: ARM: Added test infrastructure for uboot
Uboot is now running such that it can be used to download the kernel.
A binary image of u-boot is located in the kernel root directory
together with binary files (rather than elf) for the ramdisk as u-boot
needs binaries. The convert script was changed to create a binary image
of the kernel also for u-boot.