Xilinx: ARM: SMP hacks for Palladium early SMP testing
These changes create a kernel that boots on both CPUs.  The code
appears to be pretty brittle still as small changes can keep the
kernel from booting all the way and it's not clear why.

Since we don't have any control of the processor right now in
Palladium the problem is hard to debug.

This code has code that should be in a boot loader as the kernel
entry point is holding CPU1 til CPU0 gets thru some initialization
and then signals to let CPU1 go into the kernel also.

A new kernel default configuration file was added to help with SMP
testing (xilinx_palladium_smp_defconfig). I know it's a long name
but it will go anyway one day anyway.

This code does assume a special init code that is a hacked version
of Rob Pelts.  It's not cleaned up yet.  It basically boots both
CPUs to 0x8000 with some synchronization between them for the SCU
filter initialization as that's where DDR gets valid.  Both CPUs
are going thru main.