Xilinx: ARM: Adding new defconfig for the new PEEP board (EP107)
For the transition from the Dragonfire EP platform to the new PEEP
platform referred to as EP107, this should make it clearer. It can
go away at some point soon.

The new file, xilinx_ep107_defconfig, also minimizes the kernel
size right now as the platform is really slow. All kernel debug
is turned off to try to help speed it up.