Xilinx: ARM: Moving the kernel back to address 0 for PEEP2
We figured out the issue with DDR such that now it can be mapped to
0 such that the kernel can be mapped at 0.  Mapping it at 0 is easier
because it's a direct mapping between virtual and physical (mask the
top nibble from a virtual to get a physical) which makes it easier
to deal with in code.

The SMP code broke when the kernel was moved to 0x208000 because the
key address virtual to physical was no longer a direct mapping.

The following TCL commands should be used prior to loading the kernel
on the PEEP2 design so that DDR is mapped at 0, otherwise the kernel
loading will cause an error when it's most of the way loaded.

mwr 0xfef00000 0
mwr 0xfef00040 0
mwr 0xfef00000 0x2