Xilinx: ARM: Fixing SMP boot on PEEP2
The method used for SMP boot on Palladium didn't work correctly on
PEEP2. When CPU1 was polling for the boot key it seemed to stop
CPU0 from executing the kernel almost like it was being starved of
any throughput on the bus.

A WFE is added to head.S for putting CPU1 into a sleep mode so that
it isn't hogging the bus and CPU0 can boot. It was also necessary
to put an sev instruction into smp.c for CPU0 to wakeup CPU1, but onluy
after writing the key into the boot lock and flushing the caches.

Note it was necessary to put the WFE in CPU1 code in the loop that checks
the boot lock for the key because debug events (most commands in the
JTAG TCL) cause the WFE to wakeup falsely.  It's also necessary for CPU1
to write the key and flush BEFORE doing the sev instruction.