Xilinx: ARM: Prevent timer warnings due to local timers not working
This commit adds conditional compilation to the smp code around the
percpu_timer_init calls so that when local times are disabled, these
functions don't get called.  Without this, the timer was giving warnings
when the kernel booted that indicated the clockevents could not switch
to one-shot mode: dummy timer is not functional.

Once local timers are fixed, this code should be removed.