Xilinx: ARM: Updated SMP boot processing to remove hacks
This change cleans up the SMP boot processing and makes it closely
match what it should be based on the ARM model. It does still
incorporate some code that should be moved to a boot loader.

It now uses fixed addresses for the handshaking to the boot loader
code with 2 addresses, one for the address to jump to from the boot
loader, and another to signal when CPU0 is ready for CPU1 to start
the kernel.