Xilinx: ARM: Adding tcl to load the PEEP easier
There are a couple things that have to be done before loading the
kernel and it's easier to do that in tcl.

This includes making the on board ram visible at low and high
addresses rather than ROM and mapping DDR down to address 0.

Once in TCL, source the scripts to load the kernel and ramdisk.