Xilinx: ARM: fixing broadcast timer device when no local timers
The initialization of percpu timers was commented out during
previous local timer testing, but this won't work for when
local timers are disabled.  The broadcast method when local
timers are disabled require this initialization.

This appears to fix the timing problems I was seeing in the SMP
mode with processes on CPU1 not working with any time requirements.
The sleep command and top both work now on on CPUs.