Xilinx: ARM: MTD: fixing CFI for Numonyx CFI bug
The M29EW devices seem to report the CFI information wrong when
it's in 8 bit mode.

There's an app note from Numonyx on this issue and there's a patch
in the open source, but it doesn't seem to be in the mainline.

This commit is based on that patch from Massimo Cirillo titled:
buffer size adjustment for M29EW Numonyx devices.