Xilinx: ARM: Update SMP boot to work with u-boot and boot rom
New code was added to reset CPU1 to get CPU1 out of the wfe state
that the boot rom put it into.  This assumes that u-boot or something
like jtag_tcl has put a branch to 0x8000 in address 0 so that CPU1
has a boot vector to get to the kernel.

Without u-boot, a 0xE3a0F902 can be written into address 0 so that
the reset of CPU1 will work.