Xilinx: ARM: Remove GEM TX bottom-half. This permits correct SMP operation.
Remove GEM TX bottom-half.  TX processing is not so onerous that it require
a bottom half.

TX completion is now performed directly within the interrupt handler, which
can only execute on on processor at a time. No new packet processing locks
are required as NAPI poll routines are guaranteed to no be run on more than one
CPU at the same instant, and other pre-existing locks are sufficient.