Xilinx: ARM: Updated non-SMP kernel defconfigs to not use high res timers
The timer code was updated to support one-shot which allows high res timers
to work.  But high res timers require the HZ to be higher than the 10 that
we are currently using.

Values of 60 and up seem to work for non-SMP, but SMP seems to need higher than
that.  The patch for the timer one-shot mode can cause the non-SMP kernel
to not boot due to this problem (or hang while booting).

We talked about it and think that non-high res is probably the right default
for now.  The SMP kernel defaults to non-high res timers.