ARM: Xilinx: SDIO: Add platform device info
Add the platform device info for the SDIO controllers in Pele.
These SDIO controllers use the generic SDIO host controller
driver (sdhci.c). This patch also hacks the sdhci.c driver
to not check the card insertion status since this status is
not being reflected in the SDIO controllers on PEEP.

Signed-off-by: Sadanand M <sadanan@xilinx.com>