Xilinx: Axi Ethernet Driver: New Ethernet driver
This is the new Ethernet driver for Vertex6 and Spartan6
systems. The driver is based on existing LL Temac driver and
supports all features already supported by it. The driver
implements standard net device and MDIO access APIs. This driver
implements partial checksum offload feature on Tx and Rx path.

Signed-off-by: Anirudha Sarangi anirudh@xilinx.com