Xilinx: ARM: QSPI driver: Use 128 byte pages in MTD driver
There seems to be a problem in the QSPI because of which
writes larger than FIFO size are causing a problem (the
last word in some pages is corrupt). This patch hacks the
m25p80.c driver to use a smaller page size so that there
is no issue with page writes.

Signed-off-by: Sadanand M <sadanan@xilinx.com>