Xilinx: MicroBlaze: initial support for Beta of Little Endian Linux
With this commit, the user can build a little endian kernel than
should run on a SP605 reference system that is not exactly the 12.4
system to be released by the Apps group.

It has to have 64 bit multiplier and the frequency downed to 50 MHz
to get it to meet timing.