Xilinx: Kernel hack: Work around Microblaze tools issue, remove sched_clock()
This is a real hack that's temporary as we need a new version of Microblaze
tools built which will build the kernel without an error.

The following error was being seen when building the 2.6.36/37 kernel.

kernel/built-in.o: In function `sched_clock_cpu':
: relocation truncated to fit: R_MICROBLAZE_32_PCREL_LO against symbol `sched_clock'
	defined in .text section in arch/microblaze/kernel/built-in.o
make: *** [.tmp_vmlinux1] Error 1

This commit takes out the sched_clock function in the generic part of the
kernel.  Microblaze and PPC both have their own version so it shouldn't hurt.
Once the tools are fixed, then this should be removed.

This is a problem with weak symbol linking in the tools.