Xilinx: ARM: 1st commit for AMP with both CPUs
This is the 1st commit to move towards an AMP solution with Linux
running at the same time on both CPUs.

It only does some infrastructure changes and moves the address
for CPU1 testing up high to match where CPU1 would run with
the kernel running on both CPUs in AMP mode.