Xilinx: ARM: Next phase of AMP changes for UART and Timer
This is the next commit in making AMP work.  It's still not done as
this only makes CPU1 run fine on the 2nd set of devices including
the 2nd UART and Timer Counter.

There seems to be a bug in the UART driver it appears as CPU1
won't work on the 2nd UART if the 1st UART is in the device.c
file.

To make CPU1 kernel work, it must be loaded into memory at
0x4208000 and the ramdisk at 0x4800000 together with changing
the kernel command line to DF1 and the initrd address to 0x4800000.

At this point, the kernel is being loaded with u-boot, then use
JTAG_TCL to target CPU2, set the pc to 0x4208000, the continue.