Xilinx: ARM: Update of kernel startup code for AMP mode
This commit adds code for CONFIG_XILINX_AMP_CPU0_MASTER such that
the master, CPU0 in AMP mode, handles CPU1 coming out of reset
and then transfers it to 0x4208000 where the CPU1 kernel lives.

This allows a Linux application to write an opcode to address
0 (the reset vector) then reset CPU1 so that it jumps to 0x8000
which is where the kernel for CPU0 lives.

This is only a temporary hack til we have the real multicore
code in the boot rom which will do handshaking, also needed
for SMP.