Xilinx: ARM: Removed hacks for controlling CPU1 from CPU0
The hacks are only ifdefed out for now as testing is still in
progress for boot rom changes to allow these to be removed.

They should be removed all the way once it's tested well.

This removes the code that causes CPU0 to write an opcode into
memory, reset CPU1 and then hold CPU1 in a pen waiting for CPU0
to wake it up to start Linux.