arm/dt: allow bootmem reservation for device tree blob and initrd
This patch adds a function, arm_reserve_devtree(), to allow the device
tree code to reserve the memory used by the FDT blob.  It also adds
early_init_dt_setup_initrd() which is the hook used by the flat device
tree code to register the region of memory occupied by an initrd.

[based on original patch written by Jeremy Kerr <jeremy.kerr@canonical.com>]
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>