arm/dt: probe for platforms via the device tree
Add a macro to define a device-tree-probable machine
(DT_MACHINE_START/DT_MACHINE_END), and iterate through compiled-in
mdescs once we have the DT.  This patch creates a new function,
setup_machine_fdt() which is analogous to the setup_machine_atags()
created in the previous patch.  It does all the early setup needed to
use a device tree machine description.

[based on work originally written by Jeremy Kerr <jeremy.kerr@canonical.com>]
Signed-off-by: Grant Likely <grant.likely@secretlab.ca>