Xilinx: ARM: Updating for 256MB of DDR in Peep8 design
This change will probably make the kernel not backward compatible
for Peep7 and older designs. There's a #ifdef in hardware.h so that
the amount of memory can be changed back if needed. Passing the
amount of memory on the kernel command line may also work to let
this kernel work on older h/w, but I haven't tested that.