Xilinx: ARM: BSP: Added device tree support
This commit allows the kernel to be booted with a basic device
tree. This commit assumes it's combined with Grant's changes i
from his tree for device tree infrastructure in ARM arch.

This commit also includes a small device tree that has the UART
in it (arch/arm/mach-xilinx/xilinx.dts). The next commit is also
needed for the UART to work using the device tree.

Note that the device tree needs to be loaded at address 0 and not
in high memory (above the kernel). It will crash the kernel when
it's loaded high. The device tree file can be compiled with dtc
just as done with powerpc and then loaded into memory using u-boot
and tftp.