Xilinx: ARM: UART: Adding device tree support
This commit updates the UART driver to handle platform bus or
device tree (OF). Both as supported for now to allow testing,
but in the long run platform bus will be removed.

This commit also removed the registration of the UART devices
from the platform bus and assumes that they will be in the
device tree (removed from devices.c).