dt/arm: Save dtb in allocated buffer instead of reserved region
Simply reserving the initial location of the dtb is not sufficient to
prevent it getting corrupted by the kernel since it might be located
in an unsafe place.  For example, the interrupt vectors overlap it if
it is located at the base of ram.

This commit copies the dtb into an allocated buffer so that it is
guaranteed to be in a safe location.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>