Xilinx: ARM: DMA: Adding support for AXI DMA engines
Some lite testing has been done. Platform data was added to the
driver and the user has to set it up manually to match the
soft IP configuration (in devices.c).  There are 2 tests in
drivers/dma that can be renamed dmatest.c and then used to
test the appropriate DMA engine.