Xilinx: ARM: AXI GPIO driver: Added platform data
Added platform data to the existing XPS/AXI GPIO driver
so that it can be used with soft IP on PEEP. Users need
to edit the platform data in devices.c, based on system
config.
In alpha2_3.0 bitstreams the 2 channels of AXI GPIO are
connected in loopback. Tested the driver thru sysfs

Signed-off-by: Sadanand M <sadanan@xilinx.com>