arm/dt: Set __machine_arch_type to match selected machine_desc for dt match
When a dtb is passed to the kernel, the kernel will use it to find a
machine_desc instead of the machine type number passed in r1.
However, the machine_is_*() macros depend on the value in
__machine_arch_type in order to work correctly.

This patch makes the dt match code update __machine_arch_type to match
the selected machine_desc so that machine_is_*() still works.

Signed-off-by: Grant Likely <grant.likely@secretlab.ca>