Xilinx: ARM: Device Configuration driver
New driver for Device Configuration block in Pele. The driver implements
the security and PCAP modules of device config block. The security module
related testing for reading and writing the registers through sysfs is done.
The downloading of bitstream through PCAP is tested upto someextent. The
transfers from PCAP are happennig to fabric but the Program done indication
is not coming from fabric.

Signed-off-by: Vidhumouli H <vidhum@xilinx.com>