Xilinx: ARM: Added fixed address device tree option
This adds an option, that is default to on, to cause the kernel
to expect the device tree at a fixed address, 0x1000000, in memory.
This enables the go command in u-boot to be used rather than
bootm as bootm is pretty slow on the emulation platform.

This option must be turned off in the kernel configuration to
use the bootm command as it ignores R2 if it's on.