powerpc: simpleboot: Allow for passing an alternative DTB via r3 at boot
In certain circumstances, particularly under QEMU emulation where certain
devices are removed from the DTB, it is useful to pass a custom DTB via r3
to the simpleboot wrapper.

This fix checks to see if r3 points to a valid DTB, and uses that if so.

Otherwise, the linked-in DTB is used, as per previous behaviour.

Signed-off-by: John Williams <john.williams@petalogix.com>