Xilinx: ARM: QSPI driver: Add dual memory support
This patch adds a config item for dual memory support and updates
the driver to handle address translation required for dual memories
on seperate buses. In this mode, the size of the last mtd partition
is increased from 8MB to 24MB.
A new entry is added for dual memories in m25p80.c driver.

Signed-off-by: Sadanand M <sadanan@xilinx.com>