Xilinx: UART: driver needs to setup timeout for non-busybox init to work
The driver was not setting up the timeout such that it was relying on the
default timeout which was very long.

This was causing a problem when using any standard init such as sysvinit
or upstart as they open and close the serial port many times.  It was not
a problem when running busybox init.