Xilinx: ARM: AMP mode operation GIC driver cleanup
This change removes a lot of hard coded interrupt mangling that
was added for AMP mode of operation.  A new function was added
to allow the BSP to move interrupts from CPU0 to CPU1 for the
AMP mode. The BSP code should call the function for each interrupt
that is to be moved to CPU1.