Xilinx: ARM: UART: fixed UART0 lockup issue during resets
The new UART driver that is going into 2.6.40 in the mainline
seemed to add a new bug with the disable of irqs in the
console write function. There is some interaction between
the console polled code and the non-console interrupt
code.

Since the startup and shutdown functions were messing
with interrupts also but not using the port lock the driver
could get into a state such that interrupts were disabled
for UART0 and UART1 was still working because the console
polled code was only running on UART0.

DT 612129 documented the issues seen. It seemed to get much
worse on 2.6.39 but could be replicated with 2.6.38 and the
new driver with enough time. 9600 baud also seemed to find
the problem much sooner than 115k.

This will need need to be patched in the mainline also.