Xilinx: ARM: SPI: fixed non-device tree operation
After the device tree changes, the platform bus was not
working. This also has some general cleanup of the
driver name to be more consistent and to make the
device tree file easier to edit (tabs).

This also removes device tree specific code for the
base address of the device and the interrupt as it's
handled automatically now.