Xilinx: ARM: Adding DMA zone to avoid DMA into lower 512K
The lower 512K of memory can not be DMAed into. A new
memory zone is added with a hole in the lower memory to
avoid this memory.

We were previously able to see an issue with USB when copying
a large file to the USB memory stick and some debug showed
that some buffers were coming from lower memory. When the
kernel releases init memory which is low this is possible.