Xilinx: ARM: Updated xilinx_qemu_defconfig for device tree
xilinx_qemu_defconfig has been updated with device tree enabled.
It should be used to configure the kernel intended to be run on
QEMU after updating QEMU to support device tree enabled kernel.
Despite the fact that at this point not all device drivers have
been upgraded to use devcie tree, it looks like by using this
file to configure the kernek, there is no need anymore to change
anything in devcies.c file before building the kernel that is
intended to be run on QEMU.