Xilinx: ARM: BSP: Fixed SMP boot for 3.0 kernel
For the mach-zynq BSP, the 3.0 kernel requires a
call to setup the gic as the method to handle the
cross cpu calls and this was not the case previously.
Without this setup, the kernel crashes when trying
to boot the 2nd CPU.

Some other minor cleanup also as smp.h is no longer
needed.