Xilinx: ARM: BSP: Update of UART for compressed image
This change is for mach-zynq as this was not a problem in
mach-xilinx.

In QEMU, u-boot is not used and the kernel has to initialize
the UART to get any kernel output. The kernel assumes a boot
loader initializes the UART which won't work in QEMU or when
not using u-boot with a Xilinx JTAG probe.