Xilinx: ARM: BSP: Kconfig update for mach-xilinx
It seems that after moving to mach-zynq, you couldn't
configure the kernel with xilinx_defconfig, then go back
to mach-xilinx just by selecting the system type as it
would create a .config that caused compile errors.

This is a minor change that turns on V7 for the CPU for
ARCH_XILINX so you can go back as I don't want to break
all the old stuff just because we moved to the new.