Xilinx: ARM: BSP: Updated timer to use device tree
The timer was using hard coded input frequencies and this won't
work on the real chip. It now pulls it's configuration information
from the device tree including the input clock frequency.

This change also allows some backward compatability when the
timer information is not present in the device tree and this should
be removed in the future.

The early vm settings for the timer were also removed as the driver
now uses ioremap. The device tree source and binary were updated to
match.

Conflicts:

	arch/arm/boot/dts/zynq-ep107.dts
	arch/arm/mach-zynq/timer.c
	zynq-ep107.dtb

remove dtb