Xilinx: ARM: adding device tree for ZC770 with DC1
This device supports the tested devices, SD, USB, Ethernet,
UART, Timer Counter, GIC. There are still some work arounds
for these devices so that u-boot is needed to make them
work.

1. Ethernet phy not updated in Linux yet, u-boot touches
the phy and makes it work.
2. The UART driver in Linux needs to clear any pending
interrupts before enabling them as the Boot ROM is causing
a bunch. This can lock up the Linux boot.
3. The USB needs to be taked out of reset with GPIO7.