Xilinx: ARM: QSPI driver: Handle Rx data properly
In certain cases, Tx empty is not always Rx complete. In such cases,
Rx data is not handles correctly. This patch enables Rx not empty
interrupt, when Rx data is not received completely.
This patch also changes the frequency of m25p80 flash to 50Mhz, as
some instructions work only at that frequency.

Signed-off-by: Sadanand M <sadanan@xilinx.com>