Xilinx: ARM: MMC/SD: Added frequency to device tree for RTL 2.0
This change is from Sadanand with some tuning. In RTL 2.0 the
h/w changed so that the registers for timeout clock frequency and
the base clock freqency now read 0. The device tree needs to pass
this information in for the driver.

This solution is not necessarily the right long term solution as
it modifies the generic files and we may need xilinx specific files
to help that.

This has been tested on EP107 and on ZC770 successfully and booting
from SD with Busybox root filesystem.