Xilinx: ARM: USB: Removed the stream disable mode setting
Removed the stream disable mode setting in the driver to make
it work on Zynq silicon. The USB device controller was not
sending response to BULK IN requests that are part of same SCSI
read request of size more than 512 bytes.

This setting can be used in low bandwidth systems like PEEP but
irrespective of this setting, the driver works on PEEP also since
the TX FIFO setting is increased from 256 bytes to 512 bytes on
PEEP unlike on Zynq silicon which is 256 bytes only.