I2C: xilinx-ps: Fix hold flag handling
Set the hold flag whenever there is more data to send and clear it when sending
the last message. Otherwise STOP-START condintions will be generated instead of
a RESTART condition when sending multiple messages at once.

This was a patch from Lars that wouldn't apply so I manually added the change.

Signed-off-by: Lars-Peter Clausen <lars@metafoo.de>