Xilinx: ARM: Device tree entries updated for timer
The TTC has 3 timers which can operate on 3 different frequencies. The device
trees are updated have 3 different frequencies for 3 timers. Similarly three
interrupt numbers are added for the 3 timers in the device tree. The timer.c
is modified to use these frequencies from the device tree. The way interrupts
are read from the device tree is changed.