remoteproc: zynq: Forward IRQs to firmware if is forwarded to Linux
If module is forcly removed and firmware is not aware about it
and IPs can still generate IRQs. When modules is reloaded there
is space between request_irq and gic_set_cpu where IRQ can happen.
In that case forward it to firmware which should be able to handle it.

Firmware after load has IRQ off and do partially reinit GIC controller.

Signed-off-by: Michal Simek <monstr@monstr.eu>