Xilinx: ARM: hacked zynq defconfig for 3.5 quick build
This commit is intended to be reverted. It's only purpose
is to allow an early 3.5-rc1 kernel to build without
NAND, USB and local timer. NAND and USB are not building and
local timer was crashing. These will be fixed, but for now
this lets clock/power work be done on 3.5 kernel.